+++ packages/form/src/Select/Select.spec.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import { shallow } from 'enzyme';
+import { shallow, debug } from 'enzyme';
 import Select from "./Select";
 
 describe("Select", () => {
@@ -47,5 +47,23 @@ describe("Select", () => {
         ></Select>
       );
       expect(component.props()).toHaveProperty('data-qa', 'id-1234');
+    });
+
+    test('Should render a placeholder when it is set', () => {
+      const component = shallow(
+        <Select
+          id={"select-id"}
+          label={"Atomic Elements"}
+          options={atomicElements}
+          placeholder="placeholderValue"
+        ></Select>
+      );
+      const select = component.find('#select-id');
+      const placeholderOption = select.childAt(0);
+
+      expect(select.props().value).toBeUndefined();
+      expect(placeholderOption.text()).toBe('placeholderValue');
+      expect(placeholderOption.props().disabled).toBeTruthy();
+      expect(placeholderOption.props().selected).toBeTruthy();
     });
 });