+++ packages/autocomplete/src/models/SingleSelectionMode.spec.js
@@ -15,14 +15,14 @@ describe('Single Selection Mode', () => {
   });
 
   it('handle defaultValue', () => {
-    component.state.results = [{label: "LABEL", value: 'VALUE'}];
-    mode.handleDefaultValue('VALUE');
-    expect(component.inputField.value).toEqual('LABEL')
+    component.state.results = [{label: "label_1", value: 'value_1'}, {label: "label_2", value: 'value_2'}];
+    mode.handleDefaultValue('value_2');
+    expect(component.inputField.value).toEqual('label_2')
   });
 
   it('handle incorrect defaultValue', () => {
-    component.state.results = [{label: "LABEL", value: 'VALUE'}];
-    mode.handleDefaultValue(['VALUE', 'VALUE2']);
-    expect(component.inputField.value).toEqual('LABEL')
+    component.state.results = [{label: "label_1", value: 'value_1'}, {label: "label_2", value: 'value_2'}];
+    mode.handleDefaultValue(['value_1', 'value_2']);
+    expect(component.inputField.value).toEqual('label_1')
   });
 });