+++ CHANGELOG.md
@@ -8,6 +8,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/).
 ### Added
 - Added select disabled state
 
+### Added
+- `Select` add type
+- Added select disabled state
+- `Select` add type
+
 
 ## v4.5.0 - 03-09-2020