+++ CHANGELOG.md
@@ -14,9 +14,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/).
 - `AccordeonTab` Added an option to pass a custom toggle function so that state can be managed in parent components.
 
 ### Fixed
-- Autocomplete component now can handle default values when using async loading items
+- `Select` Fix select elements must be either controlled or uncontrolled react warning
+- `Autocomplete` component now can handle default values when using async loading items
 ### Added
-- Autocomplete supports placeholder and required properties
+- `Autocomplete` supports placeholder and required properties
 
 ## v4.5.0 - 03-09-2020