+++ CHANGELOG.md
@@ -8,7 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/).
 
 ### Fixed
 - `Autocomplete` Fixed a crash when a non-existing defaultValue is provided
-- Sing Selection + Sync Mode now selects the correct value
+- `Autocomplete` Fixed an issue where the correct value wasn't selected in combination single select and sync search mode
 
 ## v4.6.0 - 16-09-2020