+++ CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/).
 ## [Unreleased]
 
 ### Fixed
-- No more dissapearing elements when selecting item synchrounously and clearing field
+- 'Autocomplete' Fixed disappearing elements after selecting item synchronously and clearing field
 
 ## v4.6.0 - 16-09-2020